requirements.txt
Django>=6.0
django-allauth>=65.0.0
PyJWT>=2.12.0
mysqlclient
